i fixed it up to make it more organized and it create works
