+++ Pathway/app/src/main/java/com/example/pathway/MainActivity.java
@@ -8,6 +8,8 @@ import androidx.appcompat.app.AppCompatActivity;
 
 
 import android.text.TextUtils;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 
 import androidx.navigation.NavController;
@@ -17,26 +19,12 @@ import androidx.navigation.ui.NavigationUI;
 
 import com.example.pathway.databinding.ActivityMainBinding;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-import android.view.Menu;
-import android.view.MenuItem;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
-
-import java.util.HashMap;
 
 public class MainActivity extends AppCompatActivity {
 
     private AppBarConfiguration appBarConfiguration;
     private ActivityMainBinding binding;
-    private EditText editTextEmail, editTextUsername, editTextDOB, editTextPassword, editTextConfirmPassword;
-    private Button buttonCreateAccount;
     private FirebaseAuth firebaseAuth;
-    private FirebaseDatabase firebaseDatabase;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -47,71 +35,6 @@ public class MainActivity extends AppCompatActivity {
 
         setSupportActionBar(binding.toolbar);
         firebaseAuth = FirebaseAuth.getInstance();
-        firebaseDatabase = FirebaseDatabase.getInstance();
-
-        editTextEmail = findViewById(R.id.editTextEmail);
-        editTextUsername = findViewById(R.id.editTextUsername);
-        editTextDOB = findViewById(R.id.editTextDOB);
-        editTextPassword = findViewById(R.id.editTextPassword);
-        editTextConfirmPassword = findViewById(R.id.editTextConfirmPassword);
-
-        buttonCreateAccount = findViewById(R.id.buttonCreateAccount);
-        buttonCreateAccount.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String email = editTextEmail.getText().toString().trim();
-                String username = editTextUsername.getText().toString().trim();
-                String dob = editTextDOB.getText().toString().trim();
-                String password = editTextPassword.getText().toString().trim();
-                String confirmPassword = editTextConfirmPassword.getText().toString().trim();
-                if (TextUtils.isEmpty(email)) {
-                    Toast.makeText(getApplicationContext(), "Please enter email", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-
-                if (TextUtils.isEmpty(username)) {
-                    Toast.makeText(getApplicationContext(), "Please enter username", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-
-                if (TextUtils.isEmpty(dob)) {
-                    Toast.makeText(getApplicationContext(), "Please enter date of birth", Toast.LENGTH_LONG).show();
-                    return;
-                }
-
-                if (TextUtils.isEmpty(password)) {
-                    Toast.makeText(getApplicationContext(), "Please enter password", Toast.LENGTH_LONG).show();
-                    return;
-                }
-
-                if (TextUtils.isEmpty(confirmPassword)) {
-                    Toast.makeText(getApplicationContext(), "Please re-enter password", Toast.LENGTH_LONG).show();
-                    return;
-                }
-
-                if (!password.equals(confirmPassword)) {
-                    Toast.makeText(getApplicationContext(), "Passwords do not match", Toast.LENGTH_LONG).show();
-                    return;
-                }
-
-                firebaseAuth.createUserWithEmailAndPassword(email, password)
-                        .addOnCompleteListener(MainActivity.this, task -> {
-                            if (task.isSuccessful()) {
-                                FirebaseUser user = firebaseAuth.getCurrentUser();
-
-                                DatabaseReference databaseReference = firebaseDatabase.getReference().child("users").child(user.getUid());
-                                HashMap<String, String> hashMap = new HashMap<>();
-                                hashMap.put("username", username);
-                                hashMap.put("dob", dob);
-                                databaseReference.setValue(hashMap);
-
-                                Toast.makeText(getApplicationContext(), "Account created successfully", Toast.LENGTH_LONG).show();
-                            } else {
-                                Toast.makeText(getApplicationContext(), "Account creation failed", Toast.LENGTH_LONG).show();
-                            }
-                        });
-            }
-        });
 
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_content_main);
         appBarConfiguration = new AppBarConfiguration.Builder(navController.getGraph()).build();
