+++ Pathway/app/src/main/res/layout/register_page_fragment.xml
@@ -70,10 +70,10 @@
         android:id="@+id/buttonCreateAccount"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginBottom="84dp"
         android:text="Create Account"
         app:layout_constraintBottom_toTopOf="@+id/button_second"
-        app:layout_constraintTop_toBottomOf="@+id/editTextConfirmPassword"
-        tools:layout_editor_absoluteX="0dp" />
+        tools:layout_editor_absoluteX="16dp" />
 
     <Button
         android:id="@+id/button_second"
